First version of ContenProvider. Read description.
Although we could see the trigger also from e.g. application sensors
database, it might be easier to have it also on our own database.

Moodtracker value is saved as real.

There is no ESM table as if we are using the ESM sensor rather than
fully self made, I think there is no need for it.